Refactor delete method argument types from int to Long
- Updated the argument type of the delete method in CartDAO and CartService from int to Long.
- Modified the parameter type in the MyBatis mapper file for the delete operation from int to Long.
- Adjusted the cartDelete method in CartManagementController to reflect these changes.

+++ src/main/java/com/shoppingmall/web/controller/cart/CartManagementController.java
@@ -24,7 +24,8 @@ public class CartManagementController {
     }
 
     @RequestMapping("/delete")
-    public String cartDelete(int cartId) {
+
+    public String cartDelete(Long cartId) {
         cartService.delete(cartId);
         return "ok";
     }
